requirements.txt
chardet==5.2.0
pillow==10.2.0
PyPDF2==2.12.1
reportlab==4.0.9
